Only build JS and CSS assets when necessary
This completes the dependency tree for the build system by only updating
JS and CSS build files when the source file has changed. Additionally,
all references to JS and CSS files in the HTML code are updated.

Note that there is currently a shortcoming in detecting when to
recompile CSS files, as any change to a Less file will result in every
CSS file being rebuilt.