Make 'design' a subdirectory of 'Documentation'
This moves 'design/*' into 'Documentation/design/*'. This is since the
design files are generally "documentation" (in the sense that they
document the design process, rather than providing any meaningful
functionality).

Note, this has the side effect that configuring the project with
./configure --disable-documentation also now disables building the
contents of the design directory. I'm leaving this as it seems fairly
sensible (if we're going to call design documentation, then let's treat
it as such).

Issue #218.