server.php: Add an API for accessing variables
We don't yet have an API for accessing _SERVER variables, the reason
being that we have full control over which variables that we
access. However, the ability to call pip_server_get() is still useful
since it returns an empty string if the variable is not set, preventing
an Exception being raised about attempting to access an invalid variable
index.

Issue #207.