test/views: Replace content tests with class tests
Testing whether a view works by simply verifying all of the returned
HTML against an expected value isn't a great test, since it doesn't
really tell us anything useful about the content that is returned, and
the test content must be updated every time *anything* changes in the
view component. Instead, we'll just test that we return a String, since
that is all that we need to know to be able to server the page.

Issue #235.