js/advanced: Only submit unique values in form
This re-imagines the way in which the advanced search form
works. Instead of relying on hashing the differences between the form's
starting value and value at submission time, we instead supply a set of
default values which can be compared against the values at submission
time and used to determine whether the user has entered any meaningful
data.