Implement search results map
This modifies the behaviour of the search model such that a search
returns a results map with the following attributes:

    {
       :query                      <query map>
       :no_of_records              <no of records in database>
       :no_of_matches              <no of records which matched query>
       :no_of_returned_records     <no of records returned to user>
       :max_no_of_returned_records <max no of records returned for a query>
       :records                    <vector of record maps>
    }

This greatly simplifies the search model and controllers, allowing all
processing of the results to occur on the client side. To achieve this,
the response map is encoded in JSON format and embedded with the search
and downloads pages.

Note that at this time, the no_of_records attribute is not set
correctly.

Issue #272.